requirements.txt
requests
feedparser
yfinance
pyephem